Update: Modularise FlexibleSpacer component
