+++ src/components/alexandria/FlexibleSpacer/styles.scss
@@ -1,4 +1,4 @@
-@import '../border';
+@import 'styles/border';
 
 .flexible-spacer-component {
   border: $border-lighter;
